+++ example/lib/json_decode_service.dart
@@ -3,7 +3,10 @@
 import 'dart:async';
 import 'dart:convert' show json;
 
+import 'package:cancelation_token/cancelation_token.dart';
+import 'package:logger/web.dart';
 import 'package:squadron/squadron.dart';
+import 'package:using/using.dart';
 
 import 'json_decode_service.activator.g.dart';