+++ example/lib/json_serializable.g.dart
@@ -7,7 +7,13 @@ part of 'json_serializable.dart';
 // **************************************************************************
 
 Resource _$ResourceFromJson(Map<String, dynamic> json) =>
-    Resource(json['id'] as String, json['name'] as String);
+    $checkedCreate('Resource', json, ($checkedConvert) {
+      final val = Resource(
+        $checkedConvert('id', (v) => v as String),
+        $checkedConvert('name', (v) => v as String),
+      );
+      return val;
+    });
 
 Map<String, dynamic> _$ResourceToJson(Resource instance) => <String, dynamic>{
   'id': instance.id,
@@ -15,7 +21,13 @@ Map<String, dynamic> _$ResourceToJson(Resource instance) => <String, dynamic>{
 };
 
 ResourceError _$ResourceErrorFromJson(Map<String, dynamic> json) =>
-    ResourceError(json['type'] as String, json['message'] as String);
+    $checkedCreate('ResourceError', json, ($checkedConvert) {
+      final val = ResourceError(
+        $checkedConvert('type', (v) => v as String),
+        $checkedConvert('message', (v) => v as String),
+      );
+      return val;
+    });
 
 Map<String, dynamic> _$ResourceErrorToJson(ResourceError instance) =>
     <String, dynamic>{'type': instance.type, 'message': instance.message};