+++ chopper/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 8.5.3-dev
+
+- Add `ParameterConverter` for converting query parameters before requests are sent.
+- Improve `JsonConverter` errors for typed JSON collection responses.
+
 ## 8.5.2
 
 - Update dependencies ([#708](https://github.com/lejard-h/chopper/pull/708))